[ML] Migrate Timeseries Chart to React (#25792)
* [ML] Reactified Singleseries Chart.
* [ML] Adds an initialization test for TimeseriesChart.
* [ML] Restore original behavior of updating focus chart only when interacting with the brush selector.
* [ML] Tweaked comments.
