+++ x-pack/plugins/ml/public/timeseriesexplorer/timeseries_chart.test.js
@@ -1,0 +1,69 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+//import mockOverallSwimlaneData from './__mocks__/mock_overall_swimlane.json';
+
+import moment from 'moment-timezone';
+import { mount } from 'enzyme';
+import React from 'react';
+
+import { TimeseriesChart } from './timeseries_chart';
+
+// mocking the following files because they import some core kibana
+// code which the jest setup isn't happy with.
+jest.mock('ui/chrome', () => ({
+  getBasePath: path => path,
+  getUiSettingsClient: () => ({
+    get: jest.fn()
+  }),
+}));
+
+jest.mock('ui/time_buckets', () => ({
+  TimeBuckets: function () {
+    this.setBounds = jest.fn();
+    this.setInterval = jest.fn();
+    this.getScaledDateFormat = jest.fn();
+  }
+}));
+
+jest.mock('../services/field_format_service', () => ({
+  mlFieldFormatService: {}
+}));
+
+function getTimeseriesChartPropsMock() {
+  return {
+    contextChartSelected: jest.fn(),
+    modelPlotEnabled: false,
+    showForecast: true,
+    showModelBounds: true,
+    svgWidth: 1600,
+    timefilter: {}
+  };
+}
+
+describe('TimeseriesChart', () => {
+  const mockedGetBBox = { x: 0, y: -10, width: 40, height: 20 };
+  const originalGetBBox = SVGElement.prototype.getBBox;
+  beforeEach(() => {
+    moment.tz.setDefault('UTC');
+    SVGElement.prototype.getBBox = () => mockedGetBBox;
+  });
+  afterEach(() => {
+    moment.tz.setDefault('Browser');
+    SVGElement.prototype.getBBox = originalGetBBox;
+  });
+
+  test('Minimal initialization', () => {
+    const props = getTimeseriesChartPropsMock();
+
+    const wrapper = mount(<TimeseriesChart {...props}/>);
+
+    expect(wrapper.html()).toBe(
+      `<div class="ml-timeseries-chart-react"></div>`
+    );
+  });
+
+});
