+++ x-pack/plugins/ml/public/timeseriesexplorer/timeseriesexplorer.html
@@ -115,6 +115,7 @@
           zoom-to="zoomTo"
           auto-zoom-duration="autoZoomDuration">
         </ml-timeseries-chart>
+
       </div>
 
       <span class="panel-title euiText">
